setups/backlog-setup/src/requirements.txt
psycopg2-binary
httpx
structlog